Rename base fields to just name field
