+++ static/app/views/detectors/components/forms/common/detectorNameField.tsx
